:truck: Rename manager/compressor
