+++ .github/workflows/dockers-manager-compressor-image.yml
@@ -1,4 +1,4 @@
-name: 'Build docker image: backup-manager-compressor'
+name: 'Build docker image: manager-compressor'
 on:
   push:
     branches:
@@ -8,10 +8,10 @@ on:
       - '!internal/db/**'
       - '!internal/k8s/**'
       - 'apis/grpc/**'
-      - 'pkg/manager/backup/compressor/**'
-      - 'cmd/manager/backup/compressor/**'
+      - 'pkg/manager/compressor/**'
+      - 'cmd/manager/compressor/**'
       - 'dockers/base/Dockerfile'
-      - 'dockers/manager/backup/compressor/Dockerfile'
+      - 'dockers/manager/compressor/Dockerfile'
       - 'resources/GO_VERSION'
       - 'resources/NGT_VERSION'
   pull_request:
@@ -20,10 +20,10 @@ on:
       - '!internal/db/**'
       - '!internal/k8s/**'
       - 'apis/grpc/**'
-      - 'pkg/manager/backup/compressor/**'
-      - 'cmd/manager/backup/compressor/**'
+      - 'pkg/manager/compressor/**'
+      - 'cmd/manager/compressor/**'
       - 'dockers/base/Dockerfile'
-      - 'dockers/manager/backup/compressor/Dockerfile'
+      - 'dockers/manager/compressor/Dockerfile'
       - 'resources/GO_VERSION'
       - 'resources/NGT_VERSION'
 
@@ -37,7 +37,7 @@ jobs:
     - name: Build the Docker image
       run: |
         echo 'Build docker image'
-        make docker/build/backup-manager-compressor
+        make docker/build/manager-compressor
     - name: login to DockerHub
       run: |
         echo 'Login to DockerHub'
@@ -49,14 +49,14 @@ jobs:
       if: github.ref == 'refs/heads/master'
       run: |
         echo 'Push to DockerHub (master)'
-        imagename=`make docker/name/backup-manager-compressor`
+        imagename=`make docker/name/manager-compressor`
         docker tag ${imagename} ${imagename}:nightly
         docker push ${imagename}:nightly
     - name: push to DockerHub (pull request)
       if: github.event_name == 'pull_request'
       run: |
         echo 'Push to DockerHub (pull request)'
-        imagename=`make docker/name/backup-manager-compressor`
+        imagename=`make docker/name/manager-compressor`
         pr_num=`cat $GITHUB_EVENT_PATH | jq -r ".number"`
         docker tag ${imagename} ${imagename}:pr-${pr_num}
         docker push ${imagename}:pr-${pr_num}
@@ -64,7 +64,7 @@ jobs:
       if: startsWith( github.ref, 'refs/tags/')
       run: |
         echo 'Push to DockerHub (tags)'
-        imagename=`make docker/name/backup-manager-compressor`
+        imagename=`make docker/name/manager-compressor`
         tag_name=`echo $GITHUB_REF | sed -e 's:^refs/tags/::'`
         docker tag ${imagename} ${imagename}:${tag_name}
         docker push ${imagename}:${tag_name}
