+++ Makefile
@@ -26,7 +26,7 @@ KVS_IMAGE                       = $(NAME)-meta-redis
 NOSQL_IMAGE                     = $(NAME)-meta-cassandra
 BACKUP_MANAGER_MYSQL_IMAGE      = $(NAME)-manager-backup-mysql
 BACKUP_MANAGER_CASSANDRA_IMAGE  = $(NAME)-manager-backup-cassandra
-BACKUP_MANAGER_COMPRESSOR_IMAGE = $(NAME)-manager-backup-compressor
+MANAGER_COMPRESSOR_IMAGE        = $(NAME)-manager-compressor
 
 NGT_VERSION := $(shell cat versions/NGT_VERSION)
 NGT_REPO = github.com/yahoojapan/NGT
