+++ Makefile.d/docker.mk
@@ -24,7 +24,7 @@ docker/build: \
 	docker/build/meta-cassandra \
 	docker/build/backup-manager-mysql \
 	docker/build/backup-manager-cassandra \
-	docker/build/backup-manager-compressor
+	docker/build/manager-compressor
 
 .PHONY: docker/name/base
 docker/name/base:
@@ -98,11 +98,11 @@ docker/name/backup-manager-cassandra:
 docker/build/backup-manager-cassandra: docker/build/base
 	docker build -f dockers/manager/backup/cassandra/Dockerfile -t $(REPO)/$(BACKUP_MANAGER_CASSANDRA_IMAGE) .
 
-.PHONY: docker/name/backup-manager-compressor
-docker/name/backup-manager-compressor:
-	@echo "$(REPO)/$(BACKUP_MANAGER_COMPRESSOR_IMAGE)"
+.PHONY: docker/name/manager-compressor
+docker/name/manager-compressor:
+	@echo "$(REPO)/$(MANAGER_COMPRESSOR_IMAGE)"
 
-.PHONY: docker/build/backup-manager-compressor
-## build backup-manager-compress image
-docker/build/backup-manager-compressor: docker/build/base
-	docker build -f dockers/manager/backup/compressor/Dockerfile -t $(REPO)/$(BACKUP_MANAGER_COMPRESSOR_IMAGE) .
+.PHONY: docker/build/manager-compressor
+## build manager-compressor image
+docker/build/manager-compressor: docker/build/base
+	docker build -f dockers/manager/compressor/Dockerfile -t $(REPO)/$(MANAGER_COMPRESSOR_IMAGE) .
