+++ apis/docs/manager/compressor/index.html
@@ -175,14 +175,14 @@
         
           
           <li>
-            <a href="#backup%2fbackup_manager.proto">backup/backup_manager.proto</a>
+            <a href="#compressor%2fcompressor.proto">compressor/compressor.proto</a>
             <ul>
               
               
               
               
                 <li>
-                  <a href="#backup_manager.Backup"><span class="badge">S</span>Backup</a>
+                  <a href="#compressor.Backup"><span class="badge">S</span>Backup</a>
                 </li>
               
             </ul>
@@ -195,7 +195,7 @@
     
       
       <div class="file-heading">
-        <h2 id="backup/backup_manager.proto">backup/backup_manager.proto</h2><a href="#title">Top</a>
+        <h2 id="compressor/compressor.proto">compressor/compressor.proto</h2><a href="#title">Top</a>
       </div>
       <p></p>
 
@@ -206,7 +206,7 @@
       
 
       
-        <h3 id="backup_manager.Backup">Backup</h3>
+        <h3 id="compressor.Backup">Backup</h3>
         <p></p>
         <table class="enum-table">
           <thead>
@@ -217,7 +217,7 @@
               <tr>
                 <td>GetVector</td>
                 <td><a href="#payload.Backup.GetVector.Request">.payload.Backup.GetVector.Request</a></td>
-                <td><a href="#payload.Backup.Compressed.MetaVector">.payload.Backup.Compressed.MetaVector</a></td>
+                <td><a href="#payload.Backup.MetaVector">.payload.Backup.MetaVector</a></td>
                 <td><p></p></td>
               </tr>
             
@@ -230,14 +230,14 @@
             
               <tr>
                 <td>Register</td>
-                <td><a href="#payload.Backup.Compressed.MetaVector">.payload.Backup.Compressed.MetaVector</a></td>
+                <td><a href="#payload.Backup.MetaVector">.payload.Backup.MetaVector</a></td>
                 <td><a href="#payload.Empty">.payload.Empty</a></td>
                 <td><p></p></td>
               </tr>
             
               <tr>
                 <td>RegisterMulti</td>
-                <td><a href="#payload.Backup.Compressed.MetaVectors">.payload.Backup.Compressed.MetaVectors</a></td>
+                <td><a href="#payload.Backup.MetaVectors">.payload.Backup.MetaVectors</a></td>
                 <td><a href="#payload.Empty">.payload.Empty</a></td>
                 <td><p></p></td>
               </tr>
