+++ apis/proto/manager/compressor/compressor.proto
@@ -16,12 +16,12 @@
 
 syntax = "proto3";
 
-package backup_manager;
+package compressor;
 
-option go_package = "github.com/vdaas/vald/apis/grpc/manager/backup";
+option go_package = "github.com/vdaas/vald/apis/grpc/manager/compressor";
 option java_multiple_files = true;
-option java_package = "org.vdaas.vald.manager.backup";
-option java_outer_classname = "ValdBackupManager";
+option java_package = "org.vdaas.vald.manager.compressor";
+option java_outer_classname = "ValdCompressor";
 
 import "payload.proto";
 import "google/api/annotations.proto";
@@ -30,7 +30,7 @@ import "pb/gql.proto";
 service Backup {
   option (gql.svc_type) = QUERY;
 
-  rpc GetVector(payload.Backup.GetVector.Request) returns (payload.Backup.Compressed.MetaVector) {
+  rpc GetVector(payload.Backup.GetVector.Request) returns (payload.Backup.MetaVector) {
     option (google.api.http).get = "/vector/{uuid}";
   }
 
@@ -38,18 +38,18 @@ service Backup {
     option (google.api.http).get = "/locations/{uuid}";
   }
 
-  rpc Register(payload.Backup.Compressed.MetaVector) returns (payload.Empty) {
+  rpc Register(payload.Backup.MetaVector) returns (payload.Empty) {
     option (google.api.http) = {
-      post : "/register"
-      body : "*"
+post : "/register"
+         body : "*"
     };
     option (gql.rpc_type) = MUTATION;
   }
 
-  rpc RegisterMulti(payload.Backup.Compressed.MetaVectors) returns (payload.Empty) {
+  rpc RegisterMulti(payload.Backup.MetaVectors) returns (payload.Empty) {
     option (google.api.http) = {
-      post : "/register/multi"
-      body : "*"
+post : "/register/multi"
+         body : "*"
     };
     option (gql.rpc_type) = MUTATION;
   }
@@ -61,24 +61,24 @@ service Backup {
 
   rpc RemoveMulti(payload.Backup.Remove.RequestMulti) returns (payload.Empty) {
     option (google.api.http) = {
-      post : "/delete/multi"
-      body : "*"
+post : "/delete/multi"
+         body : "*"
     };
     option (gql.rpc_type) = MUTATION;
   }
 
   rpc RegisterIPs(payload.Backup.IP.Register.Request) returns (payload.Empty) {
     option (google.api.http) = {
-      post : "/ip"
-      body : "*"
+post : "/ip"
+         body : "*"
     };
     option (gql.rpc_type) = MUTATION;
   }
 
   rpc RemoveIPs(payload.Backup.IP.Remove.Request) returns (payload.Empty) {
     option (google.api.http) = {
-      post : "/ip/delete"
-      body : "*"
+post : "/ip/delete"
+         body : "*"
     };
     option (gql.rpc_type) = MUTATION;
   }
