+++ cmd/manager/compressor/main.go
@@ -23,8 +23,8 @@ import (
 	"github.com/vdaas/vald/internal/log"
 	"github.com/vdaas/vald/internal/runner"
 	"github.com/vdaas/vald/internal/safety"
-	"github.com/vdaas/vald/pkg/manager/backup/compressor/config"
-	"github.com/vdaas/vald/pkg/manager/backup/compressor/usecase"
+	"github.com/vdaas/vald/pkg/manager/compressor/config"
+	"github.com/vdaas/vald/pkg/manager/compressor/usecase"
 )
 
 const (
@@ -38,7 +38,7 @@ func main() {
 	if err := safety.RecoverFunc(func() error {
 		return runner.Do(
 			context.Background(),
-			runner.WithName("backup-manager-compressor"),
+			runner.WithName("manager-compressor"),
 			runner.WithVersion(version, maxVersion, minVersion),
 			runner.WithConfigLoader(func(path string) (interface{}, string, error) {
 				cfg, err := config.NewConfig(path)
