+++ k8s/manager/compressor/svc.yaml
@@ -16,7 +16,7 @@
 apiVersion: v1
 kind: Service
 metadata:
-  name: manager-backup-compressor
+  name: manager-compressor
 spec:
   ports:
   - port: 8081
@@ -31,5 +31,5 @@ spec:
     name: pprof
     targetPort: 6060
   selector:
-    app: manager-backup-compressor
+    app: manager-compressor
   clusterIP: None
