+++ pkg/manager/compressor/config/config.go
