+++ pkg/manager/compressor/model/model.go
