+++ pkg/manager/compressor/service/compress.go
