+++ pkg/manager/compressor/service/compress_option.go
