+++ pkg/manager/compressor/service/doc.go
